refactor: restructure project to improve maintainability and clarity
This commit introduces a major restructuring of the project to enhance maintainability and clarity. The main changes include:

- Moved schema-related classes and utilities from `belso/schemas` to `belso/core` to centralize core functionality.
- Consolidated provider-specific code into `belso/providers` for better organization.
- Renamed `Translator` to `SchemaProcessor` to better reflect its purpose.
- Updated documentation to reflect the new structure and removed outdated references.
- Removed deprecated files and consolidated utility functions into `belso/utils`.

These changes aim to simplify the codebase, reduce redundancy, and improve the overall architecture of the project.
