+++ belso/core/field.py
@@ -1,7 +1,108 @@
-from typing import Any, Optional, Union, List, get_origin, get_args, Dict
-from belso.utils.logging import get_logger
+# belso.core.field
+
+from typing import Type, Optional, Any, List, Dict, Union, get_origin, get_args
+
+from belso.utils import get_logger
+from belso.core.schema import Schema, BaseField
 
 logger = get_logger(__name__)
+
+class NestedField(BaseField):
+    """
+    BaseField class for nested schemas.
+    Supports all advanced validation parameters passed via BaseField.\n
+    Used by:
+    - OpenAI
+    - Google
+    - Ollama
+    - Anthropic
+    - Mistral
+    - LangChain
+    - HuggingFace
+    """
+    def __init__(
+            self,
+            name: str,
+            schema: Type[Schema],
+            description: str = "",
+            required: bool = True,
+            default: Optional[Any] = None,
+            enum: Optional[List[Any]] = None,
+            range_: Optional[tuple] = None,
+            exclusive_range: Optional[tuple] = None,
+            length_range: Optional[tuple] = None,
+            items_range: Optional[tuple] = None,
+            properties_range: Optional[tuple] = None,
+            regex: Optional[str] = None,
+            multiple_of: Optional[float] = None,
+            format_: Optional[str] = None
+        ) -> None:
+        super().__init__(
+            name=name,
+            type_=dict,
+            description=description,
+            required=required,
+            default=default,
+            enum=enum,
+            range_=range_,
+            exclusive_range=exclusive_range,
+            length_range=length_range,
+            items_range=items_range,
+            properties_range=properties_range,
+            regex=regex,
+            multiple_of=multiple_of,
+            format_=format_
+        )
+        self.schema = schema
+
+class ArrayField(BaseField):
+    """
+    BaseField class for arrays of items.
+    Supports all advanced validation parameters passed via BaseField.\n
+    Used by:
+    - OpenAI
+    - Google
+    - Ollama
+    - Anthropic
+    - Mistral
+    - LangChain
+    - HuggingFace
+    """
+    def __init__(
+            self,
+            name: str,
+            items_type: Type = str,
+            description: str = "",
+            required: bool = True,
+            default: Optional[Any] = None,
+            enum: Optional[List[Any]] = None,
+            range_: Optional[tuple] = None,
+            exclusive_range: Optional[tuple] = None,
+            length_range: Optional[tuple] = None,
+            items_range: Optional[tuple] = None,
+            properties_range: Optional[tuple] = None,
+            regex: Optional[str] = None,
+            multiple_of: Optional[float] = None,
+            format_: Optional[str] = None,
+            not_: Optional[Dict] = None
+        ) -> None:
+        super().__init__(
+            name=name,
+            type_=list,
+            description=description,
+            required=required,
+            default=default,
+            enum=enum,
+            range_=range_,
+            exclusive_range=exclusive_range,
+            length_range=length_range,
+            items_range=items_range,
+            properties_range=properties_range,
+            regex=regex,
+            multiple_of=multiple_of,
+            format_=format_
+        )
+        self.items_type = items_type
 
 class Field:
     """
@@ -47,9 +148,6 @@ class Field:
         ### Returns
         - `belso.schemas.BaseField` | `belso.schemas.NestedField` | `belso.schemas.ArrayField`: the created field instance.
         """
-        from belso.schemas.base import BaseField, Schema
-        from belso.schemas.nested import NestedField, ArrayField
-
         origin = get_origin(type_)
         args = get_args(type_)
