+++ belso/formats/__init__.py
@@ -1,10 +1,10 @@
-# belso.translator.serialization.__init__
+# belso.formats.__init__
 
-from belso.translator.serialization.json_format import (
+from belso.formats.json_format import (
     schema_to_json,
     json_to_schema
 )
-from belso.translator.serialization.xml_format import (
+from belso.formats.xml_format import (
     schema_to_xml,
     xml_to_schema
 )
