+++ belso/providers/anthropic.py
@@ -1,21 +1,21 @@
-# belso.translator.providers.langchain
+# belso.providers.anthropic
 
 from typing import Any, Dict, Type
 
-from belso.utils.logging import get_logger
-from belso.schemas import Schema, BaseField
-from belso.utils.schema_helpers import (
+from belso.utils import get_logger
+from belso.core import Schema, BaseField
+from belso.utils.helpers import (
     map_json_to_python_type,
     build_properties_dict,
     create_fallback_schema
 )
 
-# Get a module-specific logger
+# Replace standard logger with our custom logger
 logger = get_logger(__name__)
 
-def to_langchain(schema: Type[Schema]) -> Dict[str, Any]:
+def to_anthropic(schema: Type[Schema]) -> Dict[str, Any]:
     """
-    Convert a belso schema to LangChain format.\n
+    Convert a belso schema to Anthropic Claude format.\n
     ---
     ### Args
     - `schema` (`Type[belso.Schema]`): the belso schema to convert.\n
@@ -25,45 +25,46 @@ def to_langchain(schema: Type[Schema]) -> Dict[str, Any]:
     """
     try:
         schema_name = schema.__name__ if hasattr(schema, "__name__") else "unnamed"
-        logger.debug(f"Starting translation of schema '{schema_name}' to LangChain format...")
+        logger.debug(f"Starting translation of schema '{schema_name}' to Anthropic format...")
 
         properties = build_properties_dict(schema)
         required_fields = schema.get_required_fields()
 
         logger.debug(f"Found {len(schema.fields)} fields, {len(required_fields)} required.")
 
-        # Create the schema in LangChain format
-        langchain_schema = {
-            "title": schema_name,
+        # Create the schema
+        anthropic_schema = {
+            "$schema": "http://json-schema.org/draft-07/schema#",
             "type": "object",
             "properties": properties,
             "required": required_fields
         }
 
-        logger.debug("Successfully created LangChain schema.")
-        return langchain_schema
+        logger.debug("Successfully created Anthropic schema.")
+        return anthropic_schema
 
     except Exception as e:
-        logger.error(f"Error translating schema to LangChain format: {e}")
+        logger.error(f"Error translating schema to Anthropic format: {e}")
         logger.debug("Translation error details", exc_info=True)
         return {}
 
-def from_langchain(schema: Dict[str, Any]) -> Type[Schema]:
+
+def from_anthropic(schema: Dict[str, Any]) -> Type[Schema]:
     """
-    Convert a LangChain schema to belso format.\n
+    Convert an Anthropic schema to belso format.\n
     ---
     ### Args
-    - `schema` (`Dict[str, Any]`): the schema to convert.\n
+    - `schema` (`Dict[str, Any]`): the Anthropic schema to convert.\n
     ---
     ### Returns
     - `Type[belso.Schema]`: the converted belso schema.
     """
     try:
-        logger.debug("Starting conversion from LangChain schema to belso format...")
+        logger.debug("Starting conversion from Anthropic schema to belso format...")
 
         # Create a new Schema class
         class ConvertedSchema(Schema):
-            name = schema.get("title", "ConvertedFromLangChain")
+            name = "ConvertedFromAnthropic"
             fields = []
 
         # Extract properties
@@ -89,18 +90,18 @@ def from_langchain(schema: Dict[str, Any]) -> Type[Schema]:
             ConvertedSchema.fields.append(
                 BaseField(
                     name=name,
-                    type=field_type,
+                    type_=field_type,
                     description=description,
                     required=required,
                     default=default
                 )
             )
 
-        logger.debug(f"Successfully converted LangChain schema to belso schema with {len(ConvertedSchema.fields)} fields.")
+        logger.debug(f"Successfully converted Anthropic schema to belso schema with {len(ConvertedSchema.fields)} fields.")
         return ConvertedSchema
 
     except Exception as e:
-        logger.error(f"Error converting LangChain schema to belso format: {e}")
+        logger.error(f"Error converting Anthropic schema to belso format: {e}")
         logger.debug("Conversion error details", exc_info=True)
         # Return a minimal schema if conversion fails
         return create_fallback_schema()
