+++ belso/providers/langchain.py
@@ -1,10 +1,10 @@
-# belso.translator.providers.huggingface
+# belso.providers.langchain
 
 from typing import Any, Dict, Type
 
-from belso.utils.logging import get_logger
-from belso.schemas import Schema, BaseField
-from belso.utils.schema_helpers import (
+from belso.utils import get_logger
+from belso.core import Schema, BaseField
+from belso.utils.helpers import (
     map_json_to_python_type,
     build_properties_dict,
     create_fallback_schema
@@ -13,9 +13,9 @@ from belso.utils.schema_helpers import (
 # Get a module-specific logger
 logger = get_logger(__name__)
 
-def to_huggingface(schema: Type[Schema]) -> Dict[str, Any]:
+def to_langchain(schema: Type[Schema]) -> Dict[str, Any]:
     """
-    Convert a belso schema to Hugging Face format.\n
+    Convert a belso schema to LangChain format.\n
     ---
     ### Args
     - `schema` (`Type[belso.Schema]`): the belso schema to convert.\n
@@ -25,31 +25,32 @@ def to_huggingface(schema: Type[Schema]) -> Dict[str, Any]:
     """
     try:
         schema_name = schema.__name__ if hasattr(schema, "__name__") else "unnamed"
-        logger.debug(f"Starting translation of schema '{schema_name}' to Hugging Face format...")
+        logger.debug(f"Starting translation of schema '{schema_name}' to LangChain format...")
 
         properties = build_properties_dict(schema)
         required_fields = schema.get_required_fields()
 
         logger.debug(f"Found {len(schema.fields)} fields, {len(required_fields)} required.")
 
-        # Create the schema
-        huggingface_schema = {
+        # Create the schema in LangChain format
+        langchain_schema = {
+            "title": schema_name,
             "type": "object",
             "properties": properties,
             "required": required_fields
         }
 
-        logger.debug("Successfully created Hugging Face schema.")
-        return huggingface_schema
+        logger.debug("Successfully created LangChain schema.")
+        return langchain_schema
 
     except Exception as e:
-        logger.error(f"Error translating schema to Hugging Face format: {e}")
+        logger.error(f"Error translating schema to LangChain format: {e}")
         logger.debug("Translation error details", exc_info=True)
         return {}
 
-def from_huggingface(schema: Dict[str, Any]) -> Type[Schema]:
+def from_langchain(schema: Dict[str, Any]) -> Type[Schema]:
     """
-    Convert a Hugging Face schema to belso format.\n
+    Convert a LangChain schema to belso format.\n
     ---
     ### Args
     - `schema` (`Dict[str, Any]`): the schema to convert.\n
@@ -58,11 +59,11 @@ def from_huggingface(schema: Dict[str, Any]) -> Type[Schema]:
     - `Type[belso.Schema]`: the converted belso schema.
     """
     try:
-        logger.debug("Starting conversion from Hugging Face schema to belso format...")
+        logger.debug("Starting conversion from LangChain schema to belso format...")
 
         # Create a new Schema class
         class ConvertedSchema(Schema):
-            name = "ConvertedFromHuggingFace"
+            name = schema.get("title", "ConvertedFromLangChain")
             fields = []
 
         # Extract properties
@@ -88,18 +89,18 @@ def from_huggingface(schema: Dict[str, Any]) -> Type[Schema]:
             ConvertedSchema.fields.append(
                 BaseField(
                     name=name,
-                    type_=field_type,
+                    type=field_type,
                     description=description,
                     required=required,
                     default=default
                 )
             )
 
-        logger.debug(f"Successfully converted Hugging Face schema to belso schema with {len(ConvertedSchema.fields)} fields.")
+        logger.debug(f"Successfully converted LangChain schema to belso schema with {len(ConvertedSchema.fields)} fields.")
         return ConvertedSchema
 
     except Exception as e:
-        logger.error(f"Error converting Hugging Face schema to belso format: {e}")
+        logger.error(f"Error converting LangChain schema to belso format: {e}")
         logger.debug("Conversion error details", exc_info=True)
         # Return a minimal schema if conversion fails
         return create_fallback_schema()
