--- belso/schemas/__init__.py
@@ -1,13 +1,0 @@
-# belso.schemas.__init__
-
-from belso.schemas.factory import Field
-from belso.schemas.base import Schema, BaseField
-from belso.schemas.nested import NestedField, ArrayField
-
-__all__ = [
-    "Field",
-    "Schema",
-    "BaseField",
-    "NestedField",
-    "ArrayField",
-]
