--- belso/translator/__init__.py
@@ -1,5 +1,0 @@
-# belso.translator.__init__
-
-from belso.translator.translator import Translator
-
-__all__ = ["Translator"]
