--- belso/translator/utils/__init__.py
@@ -1,5 +1,0 @@
-# belso.translator.utils.__init__
-
-from belso.translator.utils.detecting import detect_schema_format
-
-__all__ = ["detect_schema_format"]
