+++ belso/utils/__init__.py
@@ -1,5 +1,11 @@
 # belso.utils.__init__
 
-from belso.utils.classes import PROVIDERS
+from belso.utils.logging import get_logger
+from belso.utils.constants import PROVIDERS
+from belso.utils.detecting import detect_schema_format
 
-__all__ = ["PROVIDERS"]
+__all__ = [
+    "get_logger",
+    "PROVIDERS",
+    "detect_schema_format"
+]
