--- belso/validator/__init__.py
@@ -1,5 +1,0 @@
-# belso.validator.__init__
-
-from belso.validator.validator import Validator
-
-__all__ = ["Validator"]
