+++ docs/source/_static/custom.css
@@ -1,5 +1,5 @@
 .rst-content .customtype {
-    color: #c7254e !important; /* rosso stile Sphinx */
+    color: #c7254e !important;
     font-family: monospace;
     font-weight: bold;
 }
