+++ docs/source/api/core.rst
@@ -1,0 +1,42 @@
+Core
+====
+
+Schema
+------
+
+.. autoclass:: belso.core.schema.Schema
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Base Field
+----------
+
+.. autoclass:: belso.core.schema.BaseField
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Array Field
+-----------
+
+.. autoclass:: belso.core.field.ArrayField
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Nested Field
+------------
+
+.. autoclass:: belso.core.field.NestedField
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Field
+-----
+
+.. autoclass:: belso.core.field.Field
+   :members:
+   :show-inheritance:
+   :undoc-members:
