+++ docs/source/api/formats.rst
@@ -1,0 +1,14 @@
+Formats
+=======
+
+JSON
+----
+
+.. autofunction:: belso.formats.json_format.schema_to_json
+.. autofunction:: belso.formats.json_format.json_to_schema
+
+XML
+---
+
+.. autofunction:: belso.formats.xml_format.schema_to_xml
+.. autofunction:: belso.formats.xml_format.xml_to_schema
