+++ docs/source/api/index.rst
@@ -6,7 +6,7 @@ This section provides detailed API documentation for belso.
 .. toctree::
    :maxdepth: 3
 
-   schemas
-   translator
-   validator
+   core
+   providers
+   formats
    utils
