+++ docs/source/api/providers.rst
@@ -1,0 +1,44 @@
+Providers
+==========
+
+OpenAI
+------
+
+.. autofunction:: belso.providers.to_openai
+.. autofunction:: belso.providers.from_openai
+
+Anthropic
+---------
+
+.. autofunction:: belso.providers.to_anthropic
+.. autofunction:: belso.providers.from_anthropic
+
+Google
+------
+
+.. autofunction:: belso.providers.to_google
+.. autofunction:: belso.providers.from_google
+
+Mistral
+-------
+
+.. autofunction:: belso.providers.to_mistral
+.. autofunction:: belso.providers.from_mistral
+
+Ollama
+------
+
+.. autofunction:: belso.providers.to_ollama
+.. autofunction:: belso.providers.from_ollama
+
+HuggingFace
+-----------
+
+.. autofunction:: belso.providers.to_huggingface
+.. autofunction:: belso.providers.from_huggingface
+
+LangChain
+---------
+
+.. autofunction:: belso.providers.to_langchain
+.. autofunction:: belso.providers.from_langchain
