--- docs/source/api/schemas.rst
@@ -1,36 +1,0 @@
-Schemas
-=======
-
-Base Schema
------------
-
-.. autoclass:: belso.schemas.Schema
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-.. autoclass:: belso.schemas.BaseField
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-Nested Schema
--------------
-
-.. autoclass:: belso.schemas.ArrayField
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-.. autoclass:: belso.schemas.NestedField
-   :members:
-   :show-inheritance:
-   :undoc-members:
-
-Factory
--------
-
-.. autoclass:: belso.schemas.Field
-   :members:
-   :show-inheritance:
-   :undoc-members:
