+++ docs/source/api/utils.rst
@@ -1,20 +1,42 @@
 Utilities
 =========
 
-Classes
--------
+Constants
+---------
 
-.. autoclass:: belso.utils.PROVIDERS
+PROVIDERS
+~~~~~~~~~
+
+.. autoclass:: belso.utils.constants.PROVIDERS
    :members:
    :show-inheritance:
    :undoc-members:
-   :special-members: __init__
 
-Schema Helpers
+Logging
+-------
+
+.. autofunction:: belso.utils.logging.get_logger
+
+Detecting
+---------
+
+.. autofunction:: belso.utils.detecting.detect_schema_format
+
+Helpers
 --------------
 
-.. autofunction:: belso.utils.schema_helpers.map_python_to_json_type
-.. autofunction:: belso.utils.schema_helpers.map_json_to_python_type
-.. autofunction:: belso.utils.schema_helpers.build_properties_dict
-.. autofunction:: belso.utils.schema_helpers.is_schema_supported
-.. autofunction:: belso.utils.schema_helpers.create_fallback_schema
+Mapping
+~~~~~~~
+
+.. autofunction:: belso.utils.helpers.map_python_to_json_type
+.. autofunction:: belso.utils.helpers.map_json_to_python_type
+
+Properties
+~~~~~~~~~~
+
+.. autofunction:: belso.utils.helpers.build_properties_dict
+
+Fallback
+~~~~~~~~
+
+.. autofunction:: belso.utils.helpers.create_fallback_schema
