--- docs/source/api/validator.rst
@@ -1,7 +1,0 @@
-Validator
-=========
-
-.. autoclass:: belso.validator.Validator
-   :members:
-   :show-inheritance:
-   :undoc-members:
