+++ examples/simple/lights_control.py
@@ -2,7 +2,7 @@ import json
 import ollama
 
 from belso.utils import PROVIDERS
-from belso import Schema, Field, Translator
+from belso import Schema, Field, SchemaProcessor
 
 # Define a simple schema for light control
 class LightsControlSchema(Schema):
@@ -17,7 +17,7 @@ class LightsControlSchema(Schema):
 
 def main():
     # Convert to Ollama format
-    ollama_schema = Translator.translate(LightsControlSchema, to=PROVIDERS.OLLAMA)
+    ollama_schema = SchemaProcessor.translate(LightsControlSchema, to=PROVIDERS.OLLAMA)
 
     print("\nConverted schema to Ollama format:")
     print(json.dumps(ollama_schema, indent=4))
